Notes:
- kube-state-metrics was moved into a subchart upstream. Configuration
  overrides were kept in `prometheus/values.yaml`.
- Preserved customised scrape jobs.
- Preserved customised servicePort for Prometheus server as Grafana's
  `values.yaml` specifies Prometheus server's URL as
  `http://prometheus-server:9090`.
- Preserved customised service names. Those too are needed by Grafana.
- Kept the new `kubernetes-service-endpoints-slow` scrape job coming
  from upstream but not the example `kubernetes-pods-slow' job. When I
  tested the former job on my appliance, it collected no metrics.

# Conflicts:
#	build/kubernetes/operator/helm-charts/base/prometheus/Chart.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/OWNERS
#	build/kubernetes/operator/helm-charts/base/prometheus/README.md
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/configmaps/server.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/deployments/pushgateway.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/deployments/server.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/ingresses/alertmanager.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/ingresses/pushgateway.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/ingresses/server.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/kube-state-metrics-clusterrole.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/kube-state-metrics-deployment.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/rbac/server-clusterrole.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/rbac/server-clusterrolebinding.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/statefulsets/alertmanager.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/templates/statefulsets/server.yaml
#	build/kubernetes/operator/helm-charts/base/prometheus/values.yaml

+++ com.vmturbo.plan.orchestrator/src/main/java/com/vmturbo/plan/orchestrator/templates/TemplatesRpcService.java
@@ -1,31 +1,42 @@
 package com.vmturbo.plan.orchestrator.templates;
 
-import java.util.HashSet;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import javax.annotation.Nonnull;
+
+import com.google.common.collect.Iterators;
+
+import com.vmturbo.common.protobuf.plan.ReservationDTO;
+import com.vmturbo.plan.orchestrator.reservation.ReservationDao;
+import io.grpc.Status;
+import io.grpc.stub.StreamObserver;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jooq.exception.DataAccessException;
 
-import io.grpc.Status;
-import io.grpc.stub.StreamObserver;
-
+import com.vmturbo.common.protobuf.plan.DeploymentProfileDTO.DeploymentProfile;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.CreateTemplateRequest;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.DeleteTemplateRequest;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.DeleteTemplatesByTargetRequest;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.EditTemplateRequest;
+import com.vmturbo.common.protobuf.plan.TemplateDTO.GetHeadroomTemplateRequest;
+import com.vmturbo.common.protobuf.plan.TemplateDTO.GetHeadroomTemplateResponse;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.GetTemplateRequest;
-import com.vmturbo.common.protobuf.plan.TemplateDTO.GetTemplatesByIdsRequest;
-import com.vmturbo.common.protobuf.plan.TemplateDTO.GetTemplatesByNameRequest;
-import com.vmturbo.common.protobuf.plan.TemplateDTO.GetTemplatesByTypeRequest;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.GetTemplatesRequest;
+import com.vmturbo.common.protobuf.plan.TemplateDTO.GetTemplatesResponse;
+import com.vmturbo.common.protobuf.plan.TemplateDTO.SingleTemplateResponse;
 import com.vmturbo.common.protobuf.plan.TemplateDTO.Template;
+import com.vmturbo.common.protobuf.plan.TemplateDTO.UpdateHeadroomTemplateRequest;
+import com.vmturbo.common.protobuf.plan.TemplateDTO.UpdateHeadroomTemplateResponse;
 import com.vmturbo.common.protobuf.plan.TemplateServiceGrpc.TemplateServiceImplBase;
+import com.vmturbo.plan.orchestrator.deployment.profile.DeploymentProfileDaoImpl;
 import com.vmturbo.plan.orchestrator.plan.NoSuchObjectException;
 import com.vmturbo.plan.orchestrator.templates.exceptions.DuplicateTemplateException;
 import com.vmturbo.plan.orchestrator.templates.exceptions.IllegalTemplateOperationException;
@@ -40,8 +51,20 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
 
     private final TemplatesDao templatesDao;
 
-    public TemplatesRpcService(@Nonnull TemplatesDao templatesDao) {
+    private final DeploymentProfileDaoImpl deploymentProfileDao;
+
+    private final int templateChunkSize;
+
+    private final ReservationDao reservationDao;
+
+    TemplatesRpcService(@Nonnull final TemplatesDao templatesDao,
+                        @Nonnull final DeploymentProfileDaoImpl deploymentProfileDao,
+                        @Nonnull final ReservationDao reservationDao,
+                        final int templateChunkSize) {
         this.templatesDao = Objects.requireNonNull(templatesDao);
+        this.deploymentProfileDao = Objects.requireNonNull(deploymentProfileDao);
+        this.reservationDao = Objects.requireNonNull(reservationDao);
+        this.templateChunkSize = templateChunkSize;
     }
 
     @Override
@@ -53,6 +76,7 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
                 .withDescription("Delete templates by target ID must have an target ID").asException());
             return;
         }
+
         try {
             final List<Template> deletedTemplates = templatesDao.deleteTemplateByTargetId(request.getTargetId());
             deletedTemplates.stream().forEach(responseObserver::onNext);
@@ -66,11 +90,31 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
 
     @Override
     public void getTemplates(GetTemplatesRequest request,
-                             StreamObserver<Template> responseObserver) {
+                             StreamObserver<GetTemplatesResponse> responseObserver) {
         try {
-            for (Template template : templatesDao.getAllTemplates()) {
-                responseObserver.onNext(template);
+            final Set<Template> templates = templatesDao.getFilteredTemplates(request.getFilter());
+            final Map<Long, Set<DeploymentProfile>> profilesByTemplateId;
+            if (request.getIncludeDeploymentProfiles()) {
+                profilesByTemplateId = deploymentProfileDao.getDeploymentProfilesForTemplates(templates.stream()
+                    .map(Template::getId)
+                    .collect(Collectors.toSet()));
+            } else {
+                profilesByTemplateId = Collections.emptyMap();
             }
+
+            Iterators.partition(templates.iterator(), templateChunkSize)
+                .forEachRemaining(templateChunk -> {
+                    GetTemplatesResponse.Builder response = GetTemplatesResponse.newBuilder();
+                    templateChunk.forEach(template -> {
+                        response.addTemplates(SingleTemplateResponse.newBuilder()
+                            .setTemplate(template)
+                            .addAllDeploymentProfile(
+                                profilesByTemplateId.getOrDefault(template.getId(),
+                                    Collections.emptySet())))
+                            .build();
+                    });
+                    responseObserver.onNext(response.build());
+                });
             responseObserver.onCompleted();
         } catch (DataAccessException e) {
             responseObserver.onError(Status.INTERNAL
@@ -81,7 +125,7 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
 
     @Override
     public void getTemplate(GetTemplateRequest request,
-                            StreamObserver<Template> responseObserver) {
+                            StreamObserver<SingleTemplateResponse> responseObserver) {
         if (!request.hasTemplateId()) {
             logger.error("Missing template ID for get template.");
             responseObserver.onError(Status.INVALID_ARGUMENT
@@ -91,7 +135,15 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
         try {
             Optional<Template> templateOptional = templatesDao.getTemplate(request.getTemplateId());
             if (templateOptional.isPresent()) {
-                responseObserver.onNext(templateOptional.get());
+                final Template template = templateOptional.get();
+                final Set<DeploymentProfile> profiles =
+                    deploymentProfileDao.getDeploymentProfilesForTemplates(
+                            Collections.singleton(template.getId()))
+                        .getOrDefault(template.getId(), Collections.emptySet());
+                responseObserver.onNext(SingleTemplateResponse.newBuilder()
+                    .setTemplate(templateOptional.get())
+                    .addAllDeploymentProfile(profiles)
+                    .build());
                 responseObserver.onCompleted();
             } else {
                 responseObserver.onError(Status.NOT_FOUND
@@ -165,6 +217,13 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
                 .withDescription("Delete template must have an template ID").asException());
             return;
         }
+
+        if (templateUsedByReservation(request.getTemplateId())) {
+            responseObserver.onError(Status.FAILED_PRECONDITION
+                    .withDescription("Delete the reservations used by this template first.").asException());
+            return;
+        }
+
         try {
             final Template template = templatesDao.deleteTemplateById(request.getTemplateId());
             responseObserver.onNext(template);
@@ -184,71 +243,65 @@ public class TemplatesRpcService extends TemplateServiceImplBase {
         }
     }
 
-    @Override
-    public void getTemplatesByType(GetTemplatesByTypeRequest request,
-                                   StreamObserver<Template> responseObserver) {
-        if (!request.hasEntityType()) {
-            logger.error("Missing entity type for get template.");
-            responseObserver.onError(Status.INVALID_ARGUMENT
-                .withDescription("Get template by type must have an entity type").asException());
-            return;
+    private boolean templateUsedByReservation(long templateId) {
+        final Set<ReservationDTO.Reservation> reservationsByTemplates =
+                reservationDao.getReservationsByTemplates(Collections.singleton(templateId));
+        if (!reservationsByTemplates.isEmpty()) {
+            logger.error("This template {} is being used by reservations {}.", templateId,
+                    reservationsByTemplates.stream().map(ReservationDTO.Reservation::getName).collect(Collectors.toSet()));
+            return true;
         }
-        try {
-            for (Template template : templatesDao.getTemplatesByEntityType(request.getEntityType())) {
-                responseObserver.onNext(template);
-            }
-            responseObserver.onCompleted();
-        } catch (DataAccessException e) {
-            responseObserver.onError(Status.INTERNAL
-                .withDescription("Failed to get template by entity type " + request.getEntityType() + ".")
-                .asException());
-        }
+        return false;
     }
 
     @Override
-    public void getTemplatesByIds(GetTemplatesByIdsRequest request,
-                                  StreamObserver<Template> responseObserver) {
-        try {
-            final Set<Long> templateIds = new HashSet<>(request.getTemplateIdsList());
-            for (Template template : templatesDao.getTemplates(templateIds)) {
-                responseObserver.onNext(template);
-            }
-            responseObserver.onCompleted();
-        } catch (DataAccessException e) {
-            logger.error("Failed to find templates.", e);
-            responseObserver.onError(Status.INTERNAL
-                .withDescription("Failed to get templates.")
-                .asException());
-        }
-    }
-
-    /**
-     * Gets templates by template name.
-     *
-     * @param request The request that contains the template name.
-     * @param responseObserver response observer
-     */
-    @Override
-    public void getTemplatesByName(final GetTemplatesByNameRequest request,
-                                   final StreamObserver<Template> responseObserver) {
-        if (!request.hasTemplateName()) {
-            logger.error("Failed to get templates because template name is missing in the request.");
-            responseObserver.onError(Status.INVALID_ARGUMENT
-                    .withDescription("Get template by name must have a template name in the request.")
-                    .asException());
+    public void getHeadroomTemplateForCluster(GetHeadroomTemplateRequest request,
+                                  StreamObserver<GetHeadroomTemplateResponse> responseObserver) {
+        if (!request.hasGroupId()) {
+            final String errMsg = "Group ID is missing.";
+            logger.error(errMsg);
+            responseObserver.onError(Status.INVALID_ARGUMENT.withDescription(errMsg).asRuntimeException());
             return;
         }
 
         try {
-            for (Template template : templatesDao.getTemplatesByName(request.getTemplateName())) {
-                responseObserver.onNext(template);
-            }
+            Optional<Template> template =
+                templatesDao.getClusterHeadroomTemplateForGroup(request.getGroupId());
+            GetHeadroomTemplateResponse.Builder response = GetHeadroomTemplateResponse.newBuilder();
+            template.ifPresent(response::setHeadroomTemplate);
+            responseObserver.onNext(response.build());
             responseObserver.onCompleted();
         } catch (DataAccessException e) {
-            logger.error("Database error occurred while getting templates by name.", e);
+            logger.error("Failed to headroom because of data access exception: Group `{}`",
+                request.getGroupId(), e);
             responseObserver.onError(Status.INTERNAL
-                    .withDescription("Database error occurred while getting templates by name.")
-                    .asException());
+                .withDescription(e.getLocalizedMessage()).asException());
+        }
+    }
+
+    @Override
+    public void updateHeadroomTemplateForCluster(UpdateHeadroomTemplateRequest request,
+                                 StreamObserver<UpdateHeadroomTemplateResponse> responseObserver) {
+        if (!request.hasGroupId() || !request.hasTemplateId()) {
+            final String errMsg = "Group ID or cluster headroom template ID is missing.";
+            logger.error(errMsg);
+            responseObserver.onError(Status.INVALID_ARGUMENT.withDescription(errMsg).asRuntimeException());
+            return;
+        }
+
+        logger.info("Updating cluster headroom template ID for group `{}` with `{}`",
+            request.getGroupId(), request.getTemplateId());
+
+        try {
+            templatesDao.setOrUpdateHeadroomTemplateForCluster(request.getGroupId(),
+                request.getTemplateId());
+            responseObserver.onNext(UpdateHeadroomTemplateResponse.getDefaultInstance());
+            responseObserver.onCompleted();
+        } catch (DataAccessException e) {
+            logger.error("Failed to update group because of data access exception: Group `{}`",
+                request.getGroupId(), e);
+            responseObserver.onError(Status.INTERNAL
+                .withDescription(e.getLocalizedMessage()).asException());
         }
     }
 }
